requirements.txt
Django==2.1.3
PyJWT==1.6.4
pymongo==3.7.2
pytz==2018.7
redis==2.10.6
